% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{class}
\name{Database}
\alias{Database}
\title{A reference class to represent database connections.}
\description{
A reference class to represent database connections.
}
\section{Fields}{

\describe{
\item{\code{Driver}}{A character vector such as "ODBC Driver 17 for SQL Server" describing the database driver to use}

\item{\code{Server}}{A character vector representing the hostname of the database}

\item{\code{Schema}}{A character vector representing the schema to create the models table}

\item{\code{Database}}{A character vector representing the SQL Server database name}

\item{\code{UID}}{A character vector such as "myuser" representing the database user to use with SQL Server authentication}

\item{\code{PWD}}{A character vector representing the password of the user to use with SQL Server authentication}

\item{\code{Port}}{An integer representing the database port}
}}


