% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl.R
\name{generate_ddl_statement}
\alias{generate_ddl_statement}
\title{Generate a DDL statement for the models table}
\usage{
generate_ddl_statement(schema)
}
\arguments{
\item{schema}{A character vector representing the schema under which the table will be created}
}
\value{
A SQL string with a DDL statement
}
\description{
Generate a DDL statement for the models table
}
