% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{prepare_registry}
\alias{prepare_registry}
\title{Prepare a SQL Server Database for Model Uploads}
\usage{
prepare_registry(db_config_path = "db.yml")
}
\arguments{
\item{db_config_path}{A character vector representing the path to a YAML file with database configuration details}
}
\description{
Prepare a SQL Server Database for Model Uploads
}
\examples{
# Ensure that a db.yml file is in the current working directory with the following
# keys filled out at the top level
#
# Database: "MyDatabase"                   # The SQL Server database name where the models will be stored
# Schema: "some_schema"                    # The (unbracketed) SQL Server database schema that the model registry will be written to
# User: "my_db_user"                       # The username that will be used for SQL Server authentication
# Password: "password123"                  # The password that will be used for SQL Server authentication
# Server: "my.database.hostname"           # The hostname of the database server
# Port: 1433                               # An integer representing the port number of the database
# Driver: "ODBC Driver 17 for SQL Server"  # A string representing the driver to be used to interact with SQL Server

# Prepare the registry
prepare_registry()
}
