% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{print_header}
\alias{print_header}
\title{Print a header with the program name and version information}
\usage{
print_header()
}
\description{
Print a header with the program name and version information
}
