% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{print_manifest}
\alias{print_manifest}
\title{Print details about models to be deployed}
\usage{
print_manifest(models)
}
\arguments{
\item{models}{A vector of \code{lm} or \code{glm} model objects}
}
\description{
Print details about models to be deployed
}
