% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_config}
\alias{read_config}
\title{Read a model configuration file}
\usage{
read_config(path = "models.yml")
}
\arguments{
\item{path}{A character vector representing the path to a YAML file with model configuration details}
}
\value{
An parsed YAML object
}
\description{
Read a model configuration file
}
